requirements.txt
asgiref==3.3.1
certifi==2020.12.5
configparser==5.0.1
Django==3.1.6
djangorestframework==3.12.2
lru-dict==1.1.7
minio==6.0.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
kafka-python==2.0.2
